Convert ShippingLineDetailsViewModel to use PriceFieldFormatter
